Update blacklisted_websites.txt
I removed the site from the list https://atomicuschart.com the site is not spam. I promote it in search engines. how did he get on that list? I found a link to the list in Google. please confirm deletion. I do not want to be in any spam list